pyproject.toml
[project]
name = "neo-librus"
version = "0.1.0"
description = "A modern, AI-enhanced library management system using Django 5, Ninja, and LangGraph."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["django", "ai", "langgraph", "async", "library"]

# 🚀 Runtime Dependencies (Production)
dependencies = [
    "django>=5.1",                  # Core framework
    "django-ninja>=1.3",            # Async API
    "django-environ>=0.11",         # 12-Factor App Config
    "uvicorn[standard]>=0.30",      # ASGI Server
    "gunicorn>=23.0",               # Production Process Manager
    "pydantic>=2.8",                # Data Validation
    "pydantic-settings>=2.4",       # Settings Management
    "httpx>=0.27",                  # Async HTTP Client
    "mysqlclient>=2.1",              # MySQL database driver
    "unidecode>=1.3",               # ASCII text transliteration
    "idna>=3.0",               
    "certifi>=2020.6.20",       
    # 🧠 AI & Logic
    "langgraph>=0.2",               # Stateful AI Workflows
    "langchain>=0.2",               # AI Tooling
    "langchain-openai>=0.1",        # OpenAI Integration
    # ⚡ Background Tasks
    "arq>=0.26",                    # Async Task Queue
    "redis>=5.0",                   # Redis Client
    # 🔍 Observability
    "structlog>=24.0",              # Structured Logging
    "opentelemetry-api>=1.26",      # Tracing API
    "opentelemetry-sdk>=1.26",      # Tracing SDK
    "django-unfold>=0.78.1",
    "asgiref>=3.11",               # ASGI utilities
    # 🧩 SQL Utilities
    "sqlparse>=0.4.4",             # SQL utilities
    "django-anymail>=14.0",       
    "requests>=2.0",             
    "urllib3>=1.26",            
    "pydantic-core==2.41.5",
    "typing-extensions>=4.0",
    "typing-inspection<=0.4.2",
    "annotated-types>=0.4",
]
# 🛠 Development Dependencies (Local & CI/CD)
[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",           # Django plugin for pytest
    "pytest-asyncio>=0.23",         # Asyncio plugin for pytest
    "testcontainers>=4.8",          # Real DB/Redis in tests
    "ruff>=0.6",                    # Linter & Formatter (Fast)
    "mypy>=1.11",                   # Static Type Checker
    "django-stubs>=5.0",            # Type stubs for Django
    "types-redis>=4.6",             # Type stubs for Redis
]

# --- 🔧 Tool Configurations ---

# 1. Build System (using uv/hatchling is standard)
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# 2. Ruff (Linter & Formatter) - The "Beautiful Code" Enforcer
[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "manage.py",
]

[tool.ruff.lint]
# Enable specific rules:
# E/W: Pycodestyle errors
# F: Pyflakes
# I: Isort (Import sorting)
# B: Flake8-bugbear (Best practices)
# UP: Pyupgrade (Modern Python syntax)
# SIM: Flake8-simplify (Code simplification)
# ASYNC: Flake8-async (Async best practices)
select = ["E", "F", "W", "I", "B", "UP", "SIM", "ASYNC"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]  # Treat 'app' as local imports

# 3. Mypy (Static Type Checking) - Strict Mode
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_untyped_defs = true    # Force type hints on all functions
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "app.core.settings.local"

# 4. Pytest (Testing Framework)
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=app.core.settings.local --reuse-db"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
asyncio_mode = "auto"           # Automatically handle async tests

[tool.hatch.build.targets.wheel]
packages = ["app"]
